refactor(tool): consolidate enum option tools
- list_work_item/document_enum_options sat in separate domain modules duplicating ~60 lines of identical logic

- consolidate both into tools/enum.py behind a shared _list_enum_options helper; migrate tests, tool names unchanged
