+++ src/mcp_server_polarion/tools/work_items.py
@@ -18,7 +18,6 @@ from mcp_server_polarion.core.exceptions import (
 )
 from mcp_server_polarion.models import (
     MAX_BODY_HTML_LEN,
-    EnumOption,
     Hyperlink,
     JsonValue,
     PaginatedResult,
@@ -42,7 +41,6 @@ from mcp_server_polarion.tools._shared.helpers import (
     STANDARD_WORK_ITEM_ATTRIBUTES,
     WORK_ITEM_DETAIL_FIELDS,
     WORK_ITEM_LIST_FIELDS,
-    build_enum_option,
     compute_has_more,
     encode_path_segment,
     extract_short_id,
@@ -591,82 +589,6 @@ async def update_work_item(  # noqa: PLR0912, PLR0913, PLR0915
         current=current,
         changes=changes,
         payload_preview=None,
-    )
-
-
-@mcp.tool(
-    tags={"read"},
-    timeout=60.0,
-    annotations={"readOnlyHint": True},
-)
-async def list_work_item_enum_options(  # noqa: PLR0913
-    ctx: Context,
-    project_id: str = Field(description="Polarion project ID."),
-    field_id: str = Field(
-        description=(
-            "e.g. 'status', 'type', 'severity', 'priority', or a custom field id."
-        ),
-    ),
-    work_item_type: str = Field(
-        description="e.g. 'task', 'requirement'; '~' = type-agnostic.",
-    ),
-    page_size: int = Field(default=DEFAULT_PAGE_SIZE, ge=1, le=100),
-    page_number: int = Field(default=1, ge=1),
-) -> PaginatedResult[EnumOption]:
-    """List valid enum option ids for a work item field of a given type.
-
-    Resolve enum ids here before create_work_items / update_work_item — enums
-    are validated on write, invalid ids raise with this set. An unknown
-    work_item_type silently falls back to ~, so verify the type id first.
-    """
-    client = get_client(ctx)
-    path = (
-        f"/projects/{encode_path_segment(project_id)}"
-        f"/workitems/fields/{encode_path_segment(field_id)}"
-        "/actions/getAvailableOptions"
-    )
-    params: dict[str, str | int] = {
-        "type": work_item_type,
-        "page[size]": page_size,
-        "page[number]": page_number,
-    }
-    try:
-        response = await client.get(path, params=params)
-    except PolarionNotFoundError as exc:
-        raise ValueError(
-            f"No enum options for field '{field_id}' on work item type "
-            f"'{work_item_type}' in project '{project_id}'."
-        ) from exc
-    except PolarionAuthError as exc:
-        raise PermissionError(
-            "Cannot list work item enum options"
-            " -- check your POLARION_TOKEN permissions."
-        ) from exc
-    except PolarionError as exc:
-        raise RuntimeError(
-            f"Failed to list enum options for field '{field_id}': {exc.message}"
-        ) from exc
-
-    data = response.get("data", [])
-    items: list[EnumOption] = []
-    if isinstance(data, list):
-        for entry in data:
-            if isinstance(entry, dict):
-                items.append(build_enum_option(entry))
-
-    raw_total = extract_total_count(response)
-    total = raw_total
-    if total <= 0 and items:
-        total = (page_number - 1) * page_size + len(items)
-
-    return PaginatedResult[EnumOption](
-        items=items,
-        total_count=total,
-        page=page_number,
-        page_size=page_size,
-        has_more=compute_has_more(
-            response, raw_total, page_number, page_size, len(items)
-        ),
     )
